src/test/java/hello/helosrping/repository/MemoryMemberRepositoryTest.java
package hello.helosrping.repository;

import hello.helosrping.domain.Member;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();

    // 테스트 케이스가 하나 끝날 때 레포지토리를 지워줘야한다.
    // 테스트는 서로 의존관계 없이 설계가 되어야 한다.
    // 테스트 케이스를 먼저 작성하고 코드를 작성하는 경우도 있다.
    // 이 것을 TDD 테스트 주도 개발이라고 한다.
    @AfterEach
    public void afterEach() {
        repository.clearStore();
    }

    @Test
    public void save() {
        Member member = new Member();
        member.setName("spring");

        repository.save(member);
        Member result = repository.findById(member.getId()).get();
        assertThat(member).isEqualTo(result);
    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        Member result = repository.findByName("spring1").get();
        assertThat(result).isEqualTo(member1);

    }

    @Test
    public void findAll() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        List<Member> all = repository.findAll();
        assertThat(all.size()).isEqualTo(2);
    }

}
